move outside of class and shared methods to helpers
